# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SuuqLink marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  buyer: "buyer",
  seller: "seller",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  avatarUrl: string;
  language: string;
  currency: string;
  /** @nullable */
  vendorId?: number | null;
  createdAt: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  buyer: "buyer",
  seller: "seller",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  role: RegisterBodyRole;
}

export type SwitchRoleBodyRole =
  (typeof SwitchRoleBodyRole)[keyof typeof SwitchRoleBodyRole];

export const SwitchRoleBodyRole = {
  buyer: "buyer",
  seller: "seller",
  admin: "admin",
} as const;

export interface SwitchRoleBody {
  role: SwitchRoleBodyRole;
}

export interface Category {
  id: number;
  slug: string;
  name: string;
  icon: string;
  accentColor: string;
  productCount: number;
}

export interface Vendor {
  id: number;
  slug: string;
  name: string;
  tagline: string;
  city: string;
  country: string;
  rating: number;
  reviewCount: number;
  productCount: number;
  logoUrl: string;
  coverUrl: string;
  verified: boolean;
  badges: string[];
}

export interface Product {
  id: number;
  slug: string;
  title: string;
  price: number;
  /** @nullable */
  compareAtPrice?: number | null;
  currency: string;
  imageUrl: string;
  vendorId: number;
  vendorName: string;
  rating: number;
  reviewCount: number;
  categorySlug: string;
  categoryName: string;
  inStock: boolean;
  badges: string[];
}

export type VendorDetail = Vendor & {
  about: string;
  joinedAt: string;
  responseTimeMins: number;
  fulfillmentRate: number;
  products: Product[];
};

export interface ProductSpec {
  label: string;
  value: string;
}

export interface Review {
  id: number;
  author: string;
  rating: number;
  body: string;
  createdAt: string;
}

export type ProductDetail = Product & {
  description: string;
  gallery: string[];
  specs: ProductSpec[];
  shippingDays: number;
  vendor: Vendor;
  reviews: Review[];
};

export interface ProductListResponse {
  items: Product[];
  total: number;
}

export interface SearchSuggestions {
  terms: string[];
  categories: Category[];
  products: Product[];
}

export interface CartItem {
  id: number;
  productId: number;
  title: string;
  imageUrl: string;
  vendorName: string;
  price: number;
  currency: string;
  quantity: number;
  lineTotal: number;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  shipping: number;
  tax: number;
  total: number;
  currency: string;
  itemCount: number;
}

export interface AddCartItemBody {
  productId: number;
  quantity: number;
}

export interface UpdateCartItemBody {
  quantity: number;
}

export interface OrderItem {
  id: number;
  productId: number;
  title: string;
  imageUrl: string;
  vendorName: string;
  price: number;
  currency: string;
  quantity: number;
}

export interface OrderTimelineEvent {
  status: string;
  label: string;
  at: string;
  completed: boolean;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  placed: "placed",
  confirmed: "confirmed",
  packed: "packed",
  shipped: "shipped",
  out_for_delivery: "out_for_delivery",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface ShippingAddress {
  fullName: string;
  phone: string;
  line1: string;
  /** @nullable */
  line2?: string | null;
  city: string;
  region: string;
  country: string;
  /** @nullable */
  notes?: string | null;
}

export interface Order {
  id: number;
  reference: string;
  status: OrderStatus;
  items: OrderItem[];
  subtotal: number;
  shipping: number;
  tax: number;
  total: number;
  currency: string;
  shippingAddress: ShippingAddress;
  paymentMethod: string;
  createdAt: string;
  estimatedDelivery: string;
  timeline: OrderTimelineEvent[];
}

export type CreateOrderBodyPaymentMethod =
  (typeof CreateOrderBodyPaymentMethod)[keyof typeof CreateOrderBodyPaymentMethod];

export const CreateOrderBodyPaymentMethod = {
  mobile_money: "mobile_money",
  card: "card",
  cash_on_delivery: "cash_on_delivery",
  bank_transfer: "bank_transfer",
} as const;

export interface CreateOrderBody {
  shippingAddress: ShippingAddress;
  paymentMethod: CreateOrderBodyPaymentMethod;
}

export interface Conversation {
  id: number;
  vendorId: number;
  vendorName: string;
  vendorLogo: string;
  lastMessage: string;
  lastMessageAt: string;
  unreadCount: number;
}

export type MessageAuthor = (typeof MessageAuthor)[keyof typeof MessageAuthor];

export const MessageAuthor = {
  me: "me",
  vendor: "vendor",
} as const;

export interface Message {
  id: number;
  conversationId: number;
  author: MessageAuthor;
  body: string;
  createdAt: string;
}

export interface ConversationDetail {
  conversation: Conversation;
  messages: Message[];
}

export interface StartConversationBody {
  vendorId: number;
  /** @nullable */
  productId?: number | null;
  /** @nullable */
  message?: string | null;
}

export interface SendMessageBody {
  body: string;
}

export interface RevenuePoint {
  date: string;
  revenue: number;
  orders: number;
}

export interface StatusBucket {
  status: string;
  count: number;
}

export interface TopVendor {
  vendorId: number;
  name: string;
  revenue: number;
  orders: number;
  logoUrl: string;
}

export interface TopCategory {
  categorySlug: string;
  name: string;
  revenue: number;
  share: number;
}

export interface AdminStats {
  totalRevenue: number;
  revenueChangePct: number;
  ordersCount: number;
  ordersChangePct: number;
  activeBuyers: number;
  activeBuyersChangePct: number;
  activeVendors: number;
  activeVendorsChangePct: number;
  currency: string;
  revenueByDay: RevenuePoint[];
  ordersByStatus: StatusBucket[];
  topVendors: TopVendor[];
  topCategories: TopCategory[];
}

export type VendorAdminRowStatus =
  (typeof VendorAdminRowStatus)[keyof typeof VendorAdminRowStatus];

export const VendorAdminRowStatus = {
  active: "active",
  pending: "pending",
  suspended: "suspended",
} as const;

export interface VendorAdminRow {
  id: number;
  name: string;
  city: string;
  country: string;
  productCount: number;
  ordersCount: number;
  revenue: number;
  rating: number;
  status: VendorAdminRowStatus;
  joinedAt: string;
}

export type ActivityItemKind =
  (typeof ActivityItemKind)[keyof typeof ActivityItemKind];

export const ActivityItemKind = {
  order: "order",
  signup: "signup",
  vendor: "vendor",
  review: "review",
  message: "message",
} as const;

export interface ActivityItem {
  id: number;
  kind: ActivityItemKind;
  label: string;
  detail: string;
  at: string;
}

/**
 * Not found
 */
export type NotFoundResponse = ErrorResponse;

export type ListVendorsParams = {
  q?: string;
  featured?: boolean;
};

export type ListProductsParams = {
  q?: string;
  /**
   * Category slug
   */
  category?: string;
  vendorId?: number;
  minPrice?: number;
  maxPrice?: number;
  sort?: ListProductsSort;
  limit?: number;
};

export type ListProductsSort =
  (typeof ListProductsSort)[keyof typeof ListProductsSort];

export const ListProductsSort = {
  popular: "popular",
  newest: "newest",
  price_asc: "price_asc",
  price_desc: "price_desc",
  rating: "rating",
} as const;

export type GetSearchSuggestionsParams = {
  q?: string;
};