# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SuuqLink marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/switch-role:
    post:
      operationId: switchRole
      tags: [auth]
      summary: Switch active demo persona (buyer/seller/admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwitchRoleBody"
      responses:
        "200":
          description: Active user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /cart/items:
    post:
      operationId: addCartItem
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCartItemBody"
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/items/{id}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove cart item
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /conversations:
    get:
      operationId: listConversations
      tags: [chat]
      summary: List my conversations
      responses:
        "200":
          description: Conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: startConversation
      tags: [chat]
      summary: Start or fetch a conversation with a vendor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartConversationBody"
      responses:
        "200":
          description: Conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /conversations/{id}:
    get:
      operationId: getConversation
      tags: [chat]
      summary: Get conversation with messages
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationDetail"
        "404":
          $ref: "#/components/responses/NotFound"
  /conversations/{id}/messages:
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationDetail"
...
components:
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
    User:
      type: object
      required: [id, name, email, role, avatarUrl, language, currency, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role:
          type: string
          enum: [buyer, seller, admin]
        avatarUrl: { type: string }
        language: { type: string }
        currency: { type: string }
        vendorId:
          type: ["integer", "null"]
        createdAt: { type: string, format: date-time }
    LoginBody:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    RegisterBody:
      type: object
      required: [name, email, password, role]
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
        role:
          type: string
          enum: [buyer, seller]
    SwitchRoleBody:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [buyer, seller, admin]
...
    CartItem:
      type: object
      required: [id, productId, title, imageUrl, vendorName, price, currency, quantity, lineTotal]
      properties:
        id: { type: integer }
        productId: { type: integer }
        title: { type: string }
        imageUrl: { type: string }
        vendorName: { type: string }
        price: { type: number }
        currency: { type: string }
        quantity: { type: integer }
        lineTotal: { type: number }
    Cart:
      type: object
      required: [items, subtotal, shipping, tax, total, currency, itemCount]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal: { type: number }
        shipping: { type: number }
        tax: { type: number }
        total: { type: number }
        currency: { type: string }
        itemCount: { type: integer }
    AddCartItemBody:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: integer }
        quantity: { type: integer }
    UpdateCartItemBody:
      type: object
      required: [quantity]
      properties:
        quantity: { type: integer }
...
    Conversation:
      type: object
      required: [id, vendorId, vendorName, vendorLogo, lastMessage, lastMessageAt, unreadCount]
      properties:
        id: { type: integer }
        vendorId: { type: integer }
        vendorName: { type: string }
        vendorLogo: { type: string }
        lastMessage: { type: string }
        lastMessageAt: { type: string, format: date-time }
        unreadCount: { type: integer }
    Message:
      type: object
      required: [id, conversationId, author, body, createdAt]
      properties:
        id: { type: integer }
        conversationId: { type: integer }
        author:
          type: string
          enum: [me, vendor]
        body: { type: string }
        createdAt: { type: string, format: date-time }
    ConversationDetail:
      type: object
      required: [conversation, messages]
      properties:
        conversation:
          $ref: "#/components/schemas/Conversation"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
    StartConversationBody:
      type: object
      required: [vendorId]
      properties:
        vendorId: { type: integer }
        productId:
          type: ["integer", "null"]
        message:
          type: ["string", "null"]
    SendMessageBody:
      type: object
      required: [body]
      properties:
        body: { type: string }
...