# lib/api-zod/src/generated/types/conversationDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SuuqLink marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Conversation } from "./conversation";
import type { Message } from "./message";

export interface ConversationDetail {
  conversation: Conversation;
  messages: Message[];
}